Merge release/v0.1.0 into development.

+++ include/pfasst/pfasst.hpp
@@ -9,20 +9,24 @@
 
 using namespace std;
 
-namespace pfasst {
+namespace pfasst
+{
 
   template<typename timeT>
-  class PFASST : public Controller<timeT> {
+  class PFASST 
+    : public Controller<timeT>
+  {
 
-    void run(timeT dt, int nsteps) {
+      void run(timeT dt, size_t nsteps)
+      {
 
-    }
+      }
 
-    // leveliter cycle_down(leveliter levels, double t, double dt);
-    // leveliter cycle_up(leveliter levels, double t, double dt);
-    // leveliter cycle_bottom(leveliter levels, double t, double dt);
-    // leveliter cycle_top(leveliter levels, double t, double dt);
-    // leveliter cycle_v(leveliter levels, double t, double dt);
+      // leveliter cycle_down(leveliter levels, double t, double dt);
+      // leveliter cycle_up(leveliter levels, double t, double dt);
+      // leveliter cycle_bottom(leveliter levels, double t, double dt);
+      // leveliter cycle_top(leveliter levels, double t, double dt);
+      // leveliter cycle_v(leveliter levels, double t, double dt);
   };
 
 }
